Initial Ratings,business_id,useful stored in DB

+++ dm.rb
@@ -9,7 +9,7 @@ class Rating
   property :id            , Serial
   property :business_id   , String
   property :useful        , Integer
-  property :initial_rating , Integer 
+  property :initial_rating , Float 
   
 end
